+++ flexmeasures/api/common/utils/validators.py
@@ -296,7 +296,7 @@ def optional_user_sources_accepted(
     return wrapper
 
 
-def optional_prior_accepted(ex_post: bool = False):
+def optional_prior_accepted(ex_post: bool = False, infer_missing: bool = True):
     """Decorator which specifies that a GET or POST request accepts an optional prior.
     It parses relevant form data and sets the "prior" keyword param.
 
@@ -305,6 +305,8 @@ def optional_prior_accepted(ex_post: bool = False):
     -   This results in the filter belief_time_window = (None, prior)
 
     Optionally, an ex_post flag can be passed to the decorator to indicate that only ex-post datetimes are allowed.
+    As a useful setting (at least for POST requests), set infer_missing to True to have servers
+    (that are not in play mode) derive a prior from the server time.
     """
 
     def wrapper(fn):
@@ -330,7 +332,14 @@ def optional_prior_accepted(ex_post: bool = False):
                     if prior < knowledge_time:
                         extra_info = "Meter data can only be observed after the fact."
                         return invalid_horizon(extra_info)
+            elif (
+                infer_missing is True
+                and current_app.config.get("FLEXMEASURES_MODE", "") != "play"
+            ):
+                # A missing prior is inferred by the server (if not in play mode)
+                prior = server_now()
             else:
+                # Otherwise, a missing prior is fine (a horizon may still be inferred by the server)
                 prior = None
 
             kwargs["prior"] = prior
@@ -342,12 +351,13 @@ def optional_prior_accepted(ex_post: bool = False):
 
 
 def optional_horizon_accepted(  # noqa C901
-    ex_post: bool = False, infer_missing: bool = True
+    ex_post: bool = False,
+    infer_missing: bool = True,
+    accept_repeating_interval: bool = False,
 ):
     """Decorator which specifies that a GET or POST request accepts an optional horizon.
-    It parses relevant form data and sets the "horizon" keyword param.
-    For POST requests, the "rolling" keyword param is also set.
-    # todo: deprecate the rolling keyword param in favour of an optional "prior" parameter for POST requests
+    The horizon should be in accordance with the ISO 8601 standard.
+    It parses relevant form data and sets the "horizon" keyword param (a timedelta).
 
     Interpretation for GET requests:
     -   Denotes "at least <horizon> before the fact (positive horizon),
@@ -359,18 +369,18 @@ def optional_horizon_accepted(  # noqa C901
         or at <horizon> after the fact (negative horizon)"
     -   this results in the assignment belief_horizon = horizon
 
-    For POST requests, if no horizon is specified, it is determined by the server based on when the API endpoint was called.
-    Optionally, an ex_post flag can be passed to the decorator to indicate that only non-positive horizons are allowed.
-    Example:
+    For example:
 
         @app.route('/postMeterData')
         @optional_horizon_accepted()
         def post_meter_data(horizon):
             return 'Meter data posted'
 
-    If the message specifies a "horizon", it should be in accordance with the ISO 8601 standard.
-    If no "horizon" is specified for a POST request, it is determined by the server.
-    The play server uses 0 hours as a default horizon, while other servers derive the horizon from the server time.
+    :param ex_post:                   if True, only non-positive horizons are allowed.
+    :param infer_missing:             if True, servers that are in play mode assume that the belief_horizon of posted
+                                      values is 0 hours. This setting is meant to be used for POST requests.
+    :param accept_repeating_interval: if True, the "rolling" keyword param is also set
+                                      (this was used for POST requests before v2.0)
     """
 
     def wrapper(fn):
@@ -394,43 +404,27 @@ def optional_horizon_accepted(  # noqa C901
                     if horizon > timedelta(hours=0):
                         extra_info = "Meter data must have a zero or negative horizon to indicate observations after the fact."
                         return invalid_horizon(extra_info)
-            elif infer_missing is True:
-                # A missing horizon is only accepted if the server can infer it
-                if "start" in form and "duration" in form:
-                    start = parse_isodate_str(form["start"])
-                    duration = parse_duration(form["duration"], start)
-                    if not start:
-                        extra_info = "Cannot parse 'start' value."
-                        current_app.logger.warning(extra_info)
-                        return invalid_period(extra_info)
-                    if start.tzinfo is None:
-                        current_app.logger.warning(
-                            "Cannot parse timezone of 'start' value"
-                        )
-                        return invalid_timezone(
-                            "Start time should explicitly state a timezone."
-                        )
-                    if not duration:
-                        extra_info = "Cannot parse 'duration' value."
-                        current_app.logger.warning(extra_info)
-                        return invalid_period(extra_info)
-                    if current_app.config.get("FLEXMEASURES_MODE", "") == "play":
-                        horizon = timedelta(hours=0)
-                    else:
-                        horizon = start + duration - server_now()
-                    rolling = False
-                else:
-                    current_app.logger.warning(
-                        "Request missing both 'horizon', 'start' and 'duration'."
+                elif rolling is True and accept_repeating_interval is False:
+                    extra_info = (
+                        "API versions 2.0 and higher use regular ISO 8601 durations instead of repeating time intervals. "
+                        "For example: R/P1D should be replaced by P1D."
                     )
-                    extra_info = "Specify a 'horizon' value, or 'start' and 'duration' values so that the horizon can be inferred."
                     return invalid_horizon(extra_info)
+            elif (
+                infer_missing is True
+                and current_app.config.get("FLEXMEASURES_MODE", "") == "play"
+            ):
+                # A missing horizon is set to zero for servers in play mode
+                horizon = timedelta(hours=0)
+            elif infer_missing is True and accept_repeating_interval is True:
+                extra_info = "Missing horizons are no longer accepted for API versions below v2.0."
+                return invalid_horizon(extra_info)
             else:
-                # Otherwise, a missing horizon is fine
+                # Otherwise, a missing horizon is fine (a prior may still be inferred by the server)
                 horizon = None
 
             kwargs["horizon"] = horizon
-            if infer_missing is True:
+            if infer_missing is True and accept_repeating_interval is True:
                 kwargs["rolling"] = rolling
             return fn(*args, **kwargs)