+++ flexmeasures/api/v2_0/routes.py
@@ -11,7 +11,7 @@ from flexmeasures.api.v1_3 import implementations as v1_3_implementations
 from flexmeasures.api.v1_3 import routes as v1_3_routes
 
 from flexmeasures.api.v2_0 import flexmeasures_api as flexmeasures_api_v2_0
-from flexmeasures.api.v2_0.implementations import assets, users
+from flexmeasures.api.v2_0 import implementations as v2_0_implementations
 
 
 # The service listing for this API version (import from previous version or update if needed)
@@ -80,7 +80,7 @@ v2_0_service_listing["services"].append(
 )
 v2_0_service_listing["services"].append(
     {
-        "name": "GET /user/<id>/password-reset",
+        "name": "PATCH /user/<id>/password-reset",
         "description": "Reset a user's password.",
     },
 )
@@ -136,7 +136,7 @@ def get_assets():
     :status 401: UNAUTHORIZED
     :status 403: INVALID_SENDER
     """
-    return assets.get()
+    return v2_0_implementations.assets.get()
 
 
 @flexmeasures_api_v2_0.route("/assets", methods=["POST"])
@@ -208,7 +208,7 @@ def post_assets():
     :status 401: UNAUTHORIZED
     :status 403: INVALID_SENDER
     """
-    return assets.post()
+    return v2_0_implementations.assets.post()
 
 
 @flexmeasures_api_v2_0.route("/asset/<id>", methods=["GET"])
@@ -253,7 +253,7 @@ def get_asset(id: int):
     :status 401: UNAUTHORIZED
     :status 403: INVALID_SENDER
     """
-    return assets.fetch_one(id)
+    return v2_0_implementations.assets.fetch_one(id)
 
 
 @flexmeasures_api_v2_0.route("/asset/<id>", methods=["PATCH"])
@@ -316,7 +316,7 @@ def patch_asset(id: int):
     :status 403: INVALID_SENDER
     :status 422: UNPROCESSABLE_ENTITY
     """
-    return assets.patch(id)
+    return v2_0_implementations.assets.patch(id)
 
 
 @flexmeasures_api_v2_0.route("/asset/<id>", methods=["DELETE"])
@@ -338,7 +338,7 @@ def delete_asset(id: int):
     :status 401: UNAUTHORIZED
     :status 403: INVALID_SENDER
     """
-    return assets.delete(id)
+    return v2_0_implementations.assets.delete(id)
 
 
 @flexmeasures_api_v2_0.route("/users", methods=["GET"])
@@ -380,7 +380,7 @@ def get_users():
     :status 401: UNAUTHORIZED
     :status 403: INVALID_SENDER
     """
-    return users.get()
+    return v2_0_implementations.users.get()
 
 
 @flexmeasures_api_v2_0.route("/user/<id>", methods=["GET"])
@@ -415,7 +415,7 @@ def get_user(id: int):
     :status 401: UNAUTHORIZED
     :status 403: INVALID_SENDER
     """
-    return users.fetch_one(id)
+    return v2_0_implementations.users.fetch_one(id)
 
 
 @flexmeasures_api_v2_0.route("/user/<id>", methods=["PATCH"])
@@ -465,12 +465,12 @@ def patch_user(id: int):
     :status 403: INVALID_SENDER
     :status 422: UNPROCESSABLE_ENTITY
     """
-    return users.patch(id)
+    return v2_0_implementations.users.patch(id)
 
 
 @flexmeasures_api_v2_0.route("/user/<id>/password-reset", methods=["PATCH"])
 @auth_token_required
-# @usef_roles_accepted(*check_access(v2_0_service_listing, "GET /user/<id>"))
+# @usef_roles_accepted(*check_access(v2_0_service_listing, "PATCH /user/<id>password-reset"))
 def reset_user_password(id: int):
     """API endpoint to reset the user password. They'll get an email to choose a new password.
 
@@ -491,7 +491,7 @@ def reset_user_password(id: int):
     :status 401: UNAUTHORIZED
     :status 403: INVALID_SENDER
     """
-    return users.reset_password(id)
+    return v2_0_implementations.users.reset_password(id)
 
 
 # endpoints from earlier versions
@@ -510,18 +510,157 @@ def get_connection():
 @as_response_type("PostPriceDataResponse")
 @auth_token_required
 @usef_roles_accepted(*list_access(v2_0_service_listing, "postPriceData"))
-@append_doc_of(v1_3_routes.post_price_data)
 def post_price_data():
-    return v1_1_implementations.post_price_data_response()
+    """API endpoint to post price data.
+
+    .. :quickref: User; Upload price data to the platform
+
+    **Optional fields**
+
+    - "horizon" (see :ref:`prognoses`)
+    - "prior" (see :ref:`prognoses`)
+
+    **Example request**
+
+    This "PostPriceDataRequest" message posts prices for hourly intervals between midnight and midnight the next day
+    for the EPEX SPOT day-ahead auction.
+    The prior indicates that the prices were published at 1pm on December 31st 2020.
+
+    .. code-block:: json
+
+        {
+            "type": "PostPriceDataRequest",
+            "market": "ea1.2018-06.localhost:epex_da",
+            "values": [
+                52.37,
+                51.14,
+                49.09,
+                48.35,
+                48.47,
+                49.98,
+                58.7,
+                67.76,
+                69.21,
+                70.26,
+                70.46,
+                70,
+                70.7,
+                70.41,
+                70,
+                64.53,
+                65.92,
+                69.72,
+                70.51,
+                75.49,
+                70.35,
+                70.01,
+                66.98,
+                58.61
+            ],
+            "start": "2021-01-01T00:00:00+01:00",
+            "duration": "PT24H",
+            "prior": "2020-12-31T13:00:00+01:00",
+            "unit": "EUR/MWh"
+        }
+
+    **Example response**
+
+    This "PostPriceDataResponse" message indicates that the prices have been processed without any error.
+
+    .. sourcecode:: json
+
+        {
+            "type": "PostPriceDataResponse",
+            "status": "PROCESSED",
+            "message": "Request has been processed."
+        }
+
+    :reqheader Authorization: The authentication token
+    :reqheader Content-Type: application/json
+    :resheader Content-Type: application/json
+    :status 200: PROCESSED
+    :status 400: INVALID_DOMAIN, INVALID_MESSAGE_TYPE, INVALID_TIMEZONE, INVALID_UNIT, REQUIRED_INFO_MISSING, UNRECOGNIZED_ASSET or UNRECOGNIZED_MARKET
+    :status 401: UNAUTHORIZED
+    :status 403: INVALID_SENDER
+    :status 405: INVALID_METHOD
+
+    """
+    return v2_0_implementations.sensors.post_price_data_response()
 
 
 @flexmeasures_api_v2_0.route("/postWeatherData", methods=["POST"])
 @as_response_type("PostWeatherDataResponse")
 @auth_token_required
 @usef_roles_accepted(*list_access(v2_0_service_listing, "postWeatherData"))
-@append_doc_of(v1_3_routes.post_weather_data)
 def post_weather_data():
-    return v1_1_implementations.post_weather_data_response()
+    """API endpoint to post weather data, such as:
+
+    - "radiation" (with kW/m² as unit)
+    - "temperature" (with °C as unit)
+    - "wind_speed" (with m/s as unit)
+
+    The sensor type is part of the unique entity address for each sensor, together with the sensor's latitude and longitude.
+
+    .. :quickref: User; Upload weather data to the platform
+
+    **Optional fields**
+
+    - "horizon" (see :ref:`prognoses`)
+    - "prior" (see :ref:`prognoses`)
+
+    **Example request**
+
+    This "PostWeatherDataRequest" message posts temperature forecasts for 15-minute intervals between 3.00pm and 4.30pm
+    for a weather sensor located at latitude 33.4843866 and longitude 126.477859. The forecasts were made at noon.
+
+    .. code-block:: json
+
+        {
+            "type": "PostWeatherDataRequest",
+            "groups": [
+                {
+                    "sensor": "ea1.2018-06.localhost:temperature:33.4843866:126.477859",
+                    "values": [
+                        20.04,
+                        20.23,
+                        20.41,
+                        20.51,
+                        20.55,
+                        20.57
+                    ]
+                }
+            ],
+            "start": "2021-01-01T15:00:00+09:00",
+            "duration": "PT1H30M",
+            "prior": "2021-01-01T12:00:00+09:00",
+            "unit": "°C"
+        }
+
+    It is allowed to send higher resolutions (in this example for instance, 30 minutes) which will be upsampled.
+
+    **Example response**
+
+    This "PostWeatherDataResponse" message indicates that the forecast has been processed without any error.
+
+    .. sourcecode:: json
+
+        {
+            "type": "PostWeatherDataResponse",
+            "status": "PROCESSED",
+            "message": "Request has been processed."
+        }
+
+    :reqheader Authorization: The authentication token
+    :reqheader Content-Type: application/json
+    :resheader Content-Type: application/json
+    :status 200: PROCESSED
+    :status 400: INVALID_DOMAIN, INVALID_MESSAGE_TYPE, INVALID_TIMEZONE, INVALID_UNIT, REQUIRED_INFO_MISSING, UNRECOGNIZED_ASSET or UNRECOGNIZED_SENSOR
+    :status 401: UNAUTHORIZED
+    :status 403: INVALID_SENDER
+    :status 405: INVALID_METHOD
+
+    """
+    return v2_0_implementations.sensors.post_weather_data_response()
 
 
 @flexmeasures_api_v2_0.route("/getPrognosis", methods=["GET"])
@@ -546,9 +685,167 @@ def get_meter_data():
 @as_response_type("PostMeterDataResponse")
 @auth_token_required
 @usef_roles_accepted(*list_access(v2_0_service_listing, "postMeterData"))
-@append_doc_of(v1_3_routes.post_meter_data)
 def post_meter_data():
-    return v1_implementations.post_meter_data_response()
+    """API endpoint to post meter data.
+
+    .. :quickref: User; Upload meter data to the platform
+
+    **Optional fields**
+
+    - "horizon" (see :ref:`prognoses`)
+    - "prior" (see :ref:`prognoses`)
+
+    **Example request**
+
+    This "PostMeterDataRequest" message posts measured consumption for 15-minute intervals between 0.00am and 1.30am for
+    charging stations 1, 2 and 3 (negative values denote production).
+
+    .. code-block:: json
+
+        {
+            "type": "PostMeterDataRequest",
+            "groups": [
+                {
+                    "connections": [
+                        "CS 1",
+                        "CS 3"
+                    ],
+                    "values": [
+                        306.66,
+                        306.66,
+                        0,
+                        0,
+                        306.66,
+                        306.66
+                    ]
+                },
+                {
+                    "connections": [
+                        "CS 2"
+                    ],
+                    "values": [
+                        306.66,
+                        0,
+                        0,
+                        0,
+                        306.66,
+                        306.66
+                    ]
+                }
+            ],
+            "start": "2021-01-01T00:00:00Z",
+            "duration": "PT1H30M",
+            "unit": "MW"
+        }
+
+    It is allowed to send higher resolutions (in this example for instance, 30 minutes) which will be upsampled.
+
+    **Example response**
+
+    This "PostMeterDataResponse" message indicates that the measurement has been processed without any error.
+
+    .. sourcecode:: json
+
+        {
+            "type": "PostMeterDataResponse",
+            "status": "PROCESSED",
+            "message": "Request has been processed."
+        }
+
+    :reqheader Authorization: The authentication token
+    :reqheader Content-Type: application/json
+    :resheader Content-Type: application/json
+    :status 200: PROCESSED
+    :status 400: INVALID_DOMAIN, INVALID_MESSAGE_TYPE, INVALID_TIMEZONE, INVALID_UNIT, REQUIRED_INFO_MISSING, UNRECOGNIZED_ASSET or UNRECOGNIZED_CONNECTION_GROUP
+    :status 401: UNAUTHORIZED
+    :status 403: INVALID_SENDER
+    :status 405: INVALID_METHOD
+    """
+    return v2_0_implementations.post_meter_data_response()
+
+
+@flexmeasures_api_v2_0.route("/postPrognosis", methods=["POST"])
+@as_response_type("PostPrognosisResponse")
+@auth_token_required
+@usef_roles_accepted(*list_access(v2_0_service_listing, "postPrognosis"))
+def post_prognosis():
+    """API endpoint to post prognoses about meter data.
+
+    .. :quickref: User; Upload prognosis to the platform
+
+    **Optional fields**
+
+    - "horizon" (see :ref:`prognoses`)
+    - "prior" (see :ref:`prognoses`)
+
+    **Example request**
+
+    This "PostPrognosisRequest" message posts prognosed consumption for 15-minute intervals between 0.00am and 1.30am for
+    charging stations 1, 2 and 3 (negative values denote production), prognosed at 6pm the previous day.
+
+    .. code-block:: json
+
+        {
+            "type": "PostPrognosisRequest",
+            "groups": [
+                {
+                    "connections": [
+                        "ea1.2018-06.localhost:1:3",
+                        "ea1.2018-06.localhost:1:4"
+                    ],
+                    "values": [
+                        300,
+                        300,
+                        300,
+                        0,
+                        0,
+                        300
+                    ]
+                },
+                {
+                    "connections": [
+                        "ea1.2018-06.localhost:1:5"
+                    ],
+                    "values": [
+                        300,
+                        0,
+                        0,
+                        0,
+                        300,
+                        300
+                    ]
+                }
+            ],
+            "start": "2021-01-01T00:00:00Z",
+            "duration": "PT1H30M",
+            "prior": "2020-12-31T18:00:00Z",
+            "unit": "MW"
+        }
+
+    It is allowed to send higher resolutions (in this example for instance, 30 minutes) which will be upsampled.
+
+    **Example response**
+
+    This "PostPrognosisResponse" message indicates that the prognosis has been processed without any error.
+
+    .. sourcecode:: json
+
+        {
+            "type": "PostPrognosisResponse",
+            "status": "PROCESSED",
+            "message": "Request has been processed."
+        }
+
+    :reqheader Authorization: The authentication token
+    :reqheader Content-Type: application/json
+    :resheader Content-Type: application/json
+    :status 200: PROCESSED
+    :status 400: INVALID_MESSAGE_TYPE, INVALID_TIMEZONE, INVALID_UNIT, REQUIRED_INFO_MISSING, UNRECOGNIZED_ASSET or UNRECOGNIZED_CONNECTION_GROUP
+    :status 401: UNAUTHORIZED
+    :status 403: INVALID_SENDER
+    :status 405: INVALID_METHOD
+    """
+    return v2_0_implementations.sensors.post_prognosis_response()
 
 
 @flexmeasures_api_v2_0.route("/getService", methods=["GET"])